wallet-api: Add 'Ada' type
In preparation for the multi-currency ledger this commit adds a new type
'Ada' with operations. It also puts the 'Value' type into its own module
with a new, smaller interface.

The biggest impact from the user's point of view is that 'Value' does
not have a 'Num' instance anymore. The reason is that 'Value' is really
a vector space so there is no good definition of 'Num.fromInteger'.