contracts: Add `Ledger.Constraints`
The core of this change is the `Ledger.Constraints` module, explained
below. The rest is changing the on-chain and off-chain code to use the
constraints.

=== Ledger.Constraints.TxConstraints

Constraints on pending transactions are expressed in the `TxConstraints
i o` type. It is very similar to the old `PendingTx`, except that it has
two type parameters for the inputs and outputs. The constructors defined
in `Ledger.Constraints.TxConstraints` are also pretty much the same that
existed for `PendingTx`.

== On-chain vs off-chain

There are three incarnations of `TxConstraints`:

1. Off-chain (untyped): `type LedgerTxConstraints  = TxConstraints
[LTx.TxIn] [LTx.TxOut]`, pretty much identical to `PendingTx`. There are
two reasons we can't use the same type in on-chain code. First, the
`TxIn` transaction inputs contain validators in full (un-hashed) which
we don't want to have in the on-chain code. This will probably go away
when we update the emulator types to take the validator out of the
transaction body. Second, `TxIn` and `TxOut` only contain `Data` and
hashes. They don't say anything about the types of data and redeemer
values. This makes sense from the ledger's point of view, but when we're
inside a validator script we know the types of our own data and redeemer
values, and it would be nice to be able to write constraints using those
types directly, without having to convert from `Data` first. (That was
the motivation for the typed tx library as well)

2. On-chain: Here we have `type PendingTxConstraints a = TxConstraints
() [OnChainUtxo a]`. `a` is the type of the validator's data value. This
unlocks a new constructor `mustPayToOwnAddress :: forall i a. (Monoid i,
IsData a) => Value -> a -> TxConstraints i [OnChainUtxo a]` which we can
use to say something about the "continuing outputs" of the validator.
(Note that the idea of "own address" also doesn't exist in off-chain
code because it depends on the specific input that is being validated)

3. Off-chain (typed): `type TypedTxConstraints (ins :: [Type]) (outs ::
[Type])` is a typed interface to `TxConstraints`. It's very similar to
what we have `Ledger.Typed.Tx`.

To convert between the three representations we have the following
functions:

- `toLedgerTx :: LedgerTxConstraints -> Tx`
- `fromLedgerTx :: Tx -> LedgerTxConstraints`
- `toLedgerConstraints :: PlutusTx.IsData a => Address ->
PendingTxConstraints a -> LedgerTxConstraints`
- `toTypedTxConstraints :: IsData (DataType inn) => ScriptInstance inn
-> PendingTxConstraints (DataType inn) -> Maybe (Either
(TypedTxConstraints '[] '[]) (TypedTxConstraints '[] '[inn]))` (used for
the state machine)
- `toUntypedLedgerConstraints :: forall inn out.  TypedTxConstraints inn
out -> LedgerTxConstraints`

The `WriteTx` effect has been changed to take a `LedgerTxConstraints`
value. As a result, if we write our validator as a function that
produces `PendingTxConstraints a`, then we can use the same function to
generate the constraints that are eventually turned into a transaction
in off-chain code.

== Semantics

The semantics of the constraints are given by
`Ledger.Constraints.OnChain.checkPendingTx`

```
checkPendingTx :: (IsData a) => PendingTxConstraints a -> PendingTx ->
Bool
```

We can view `checkPendingTx` as a relation between constraints and the
pending transactions that satisfy the constraints. Also,
`checkPendingTx` relates the partial orders of `PendingTxConstraints a`
and `PendingTx` (defined by the partial orders of their constituent
parts) in a way that looks almost like a galois connection (or is that
trivial? definitely needs some more thought)

Note about ordering: The constructors for `PendingTxConstraints` cannot
express any information about the ordering of the transaction outputs,
and the actual order of the outputs of a `PendingTx` does not affect the
result of `checkPendingTx`. So we treat both inputs and outputs as if
they were (multi-) sets. IMO the fact that outputs are ordered is an
artifact of the ledger implementation, and relying on them to be in a
particular order just makes the scripts more brittle and less
composable.

`PendingTxConstraints` and `checkPendingTx` can be used in any
validator.

=== State machine

The state machine type has been changed to

```
data StateMachine s i = StateMachine {
    -- | The transition function of the state machine. 'Nothing'
    --   indicates an invalid transition from the current state.
    smTransition :: s -> i -> Value -> Maybe (PendingTxConstraints s),
    -- | The condition checking function. Checks whether a
    --   given state transition is allowed given the
    --  'PendingTx'.
    smCheck :: s -> i -> PendingTx -> Bool
    }
```

Note the transition function now takes an
extra argument (the value locked by the
output we're validating), and its result
type is `Maybe (PendingTxConstraints s)`.
The `Value` of the output is included
because it is as much a part of the
contract's state as is the data value, and
the result type combines the new state of
the machine with constraints about the
transaction that makes the transition.
`smCheck` is provided as an escape hatch in
case the constraints aren't enough. None of
our current examples need `smCheck`.
The state machine client (in
`Language.Plutus.Contract.StateMachine`)
uses the `TypedTxConstraints` type to ensure
that the machine has exactly one input and
one or zero outputs

=== Other notable changes

* `PendingTxOut` has been replaced by `TxOut` (they were identical)
* The validators increased in size (At least some of that is explained by
  additional `trace` strings. We should be able to get rid of those for
  the release version of the contracts)